# docs/openapi.yaml
basePath: /proxy/v1
definitions:
  common.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  common.PaginatedSearchResponse:
    properties:
      metadata:
        $ref: '#/definitions/common.PaginationMetadata'
      results:
        items:
          $ref: '#/definitions/models.SearchItem'
        type: array
    type: object
  common.PaginationMetadata:
    properties:
      page:
        type: integer
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
...
  models.Author:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
...
  models.GameResponse:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      description:
        type: string
      game_modes:
        items:
          type: string
        type: array
      game_type:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      image_url:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ImageEntry'
        type: array
      platforms:
        items:
          $ref: '#/definitions/models.Platform'
        type: array
      play_time:
        $ref: '#/definitions/models.PlayTime'
      release_date:
        type: string
      series:
        type: string
      themes:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  models.ImageEntry:
    properties:
      image_url:
        type: string
      size:
        $ref: '#/definitions/models.ImageSize'
      type:
        $ref: '#/definitions/models.ImageType'
    type: object
  models.ImageSize:
    enum:
    - standard
    - original
    type: string
    x-enum-varnames:
    - ImageSizeStandard
    - ImageSizeOriginal
  models.ImageType:
    enum:
    - poster
    - gallery
    type: string
    x-enum-varnames:
    - ImageTypePoster
    - ImageTypeGallery
  models.MovieResponse:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      image_url:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ImageEntry'
        type: array
      imdb_id:
        type: string
      original_title:
        type: string
      platforms:
        additionalProperties:
          items:
            $ref: '#/definitions/models.Platform'
          type: array
        type: object
      release_date:
        type: string
      status:
        type: string
      tagline:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  models.Platform:
    properties:
      image_url:
        type: string
      name:
        type: string
    type: object
  models.PlayTime:
    properties:
      completely:
        type: integer
      hastily:
        type: integer
      normally:
        type: integer
    type: object
  models.SearchItem:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      original_title:
        type: string
      release_date:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API proxy that aggregates metadata from TMDB, IGDB, Spotify, and OpenLibrary.
    Authentication: send API key via X-Api-Key header or Authorization: Bearer <token>.
    Rate limit: 60 requests/minute on protected endpoints.
  title: Denn Proxy API
  version: "1.0"
paths:
...
  /games/{id}:
    get:
      parameters:
      - description: IGDB game ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get game details
      tags:
      - Games
  /games/bulk:
    get:
      description: Fetches full details for up to 50 games in parallel.
      parameters:
      - description: Comma-separated IGDB game IDs (max 50)
        example: 7346,1020
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GameResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get multiple games by ID
      tags:
      - Games
  /games/search:
    get:
      description: 'Searches via IGDB. Note: total_pages and total_results are always
        0 (IGDB limitation).'
      parameters:
      - description: Search term
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Search games
      tags:
      - Games
  /games/trending:
    get:
      description: 'Custom trending algorithm: 70% want-to-play + 30% visits, with
        recency multiplier (up to 4x). Browser-only games excluded. total_pages and
        total_results are always 0.'
      parameters:
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get trending games
      tags:
      - Games
...
  /movies/{id}:
    get:
      parameters:
      - description: TMDB movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: ISO 3166-1 alpha-2 country code for platform availability (default
          US)
        in: header
        name: X-User-Country
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get movie details
      tags:
      - Movies
  /movies/bulk:
    get:
      description: Fetches full details for up to 50 movies in parallel. Failed IDs
        are silently omitted.
      parameters:
      - description: Comma-separated TMDB movie IDs (max 50)
        example: 550,603,680
        in: query
        name: ids
        required: true
        type: string
      - description: ISO 3166-1 alpha-2 country code for platform availability (default
          US)
        in: header
        name: X-User-Country
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MovieResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get multiple movies by ID
      tags:
      - Movies
  /movies/search:
    get:
      parameters:
      - description: Search term
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Search movies
      tags:
      - Movies
  /movies/trending:
    get:
      parameters:
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get popular/trending movies
      tags:
      - Movies
...
securityDefinitions:
  ApiKeyHeader:
    description: API key passed in the X-Api-Key header
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'API key passed as a Bearer token: "Authorization: Bearer <key>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"